fix: lock user row on buy/sell/withdraw to close race condition, use DECIMAL for money

+++ server/db/schema/schema.sql
@@ -10,7 +10,7 @@ CREATE TABLE Users (
 CREATE TABLE CashTransactions (
     cashTransactionId INTEGER AUTO_INCREMENT PRIMARY KEY,
     cashTransactionType ENUM('deposit', 'withdraw') NOT NULL,
-    amount FLOAT NOT NULL,
+    amount DECIMAL(18,8) NOT NULL,
     cashTransactionDate DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
     userId INTEGER NOT NULL,
     FOREIGN KEY (userId) REFERENCES Users(userId) ON DELETE CASCADE
@@ -20,9 +20,9 @@ CREATE TABLE AssetTransactions (
     assetTransactionId INTEGER AUTO_INCREMENT PRIMARY KEY,
     assetType ENUM('stock', 'crypto') NOT NULL,
     ticker VARCHAR(12) NOT NULL,
-    qty FLOAT NOT NULL,
-    price FLOAT NOT NULL,
-    val FLOAT NOT NULL,
+    qty DECIMAL(18,8) NOT NULL,
+    price DECIMAL(18,8) NOT NULL,
+    val DECIMAL(18,8) NOT NULL,
     assetTransactionType ENUM('buy', 'sell') NOT NULL,
     assetTransactionDate DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
     userId INTEGER NOT NULL,
